backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
supabase==2.4.0
stripe==7.0.0
PyJWT==2.10.1
python-multipart==0.0.6
httpx==0.25.2
